+++ cmd/airborne-freeze/main.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 	"time"
 
-	chassis "github.com/ai8future/chassis-go/v8"
-	"github.com/ai8future/chassis-go/v8/registry"
+	chassis "github.com/ai8future/chassis-go/v9"
+	"github.com/ai8future/chassis-go/v9/registry"
 
 	"github.com/ai8future/airborne/internal/config"
 	"github.com/ai8future/airborne/internal/tenant"